armsoc_dri2: Synchronize non-pageflip swaps to vblank as well
Wait until vblank before performing the exchange/blit for a swap. If we
do this, the X server will throttle DRI2 swaps for us.

Note that this doesn't break the swap-interval = 0 case -- our vblank
callback will simply fire immediately.

Combined with some more work in Mali to get it to stop calling WaitMSC,
this gets us perfect swap behavior for both the flip and non-flip paths.