Cleanup umplock usage
LOCK_IOCTL_CREATE is now a no-op on the kernel side, so lets not
waste time calling it.

Contrary to the way that the LOCK_IOCTL_PROCESS loop previously
appeared, this ioctl actually returns when the lock is taken,
sleeping if it needs to.

It will only return error if it is interrupted by a signal.
However we've now added SA_RESTART to X's signal handlers, so
that condition should be silently handled too.

Adjust the code to reflect the way this really works.

[endlessm/eos-shell#4869]