Create direct reference from DRI2Buffer to backing BO
When changing resolution twice in quick succession, the following
happens:

 1. bo1 is the current scanout, bo7 is the back buffer
 2. Perform first resolution change.
 3. New bo3 allocated, becomes scanout. Old bo1 scheduled for deletion.
 4. Perform second resolution change, back to original resolution.
 5. New bo8 allocated, becomes scanout. Old bo3 scheduled for deletion.
 6. ScheduleSwap happens, bo7 is flipped onto the screen.
 7. Client calls GetBuffers to get the back buffer. X has cached that bo1
    is a good match here, and returns that to the client.

 --> crash because bo1 was deleted.

It's a bit unfortunate that X is caching the old bo there. In other
circumstances it would destroy that cache in the final step, because
the resolution would not match.

Anyway, at least X is fairly honest about what it has in its caches:
we can presume that it is caching any DRI2Buffer for which it has not
called DestroyBuffer on. By adding a reference from DRI2Buffer to BO,
we can avoid freeing any object that X may still be holding onto.

[endlessm/eos-shell#5082]